🔬 convert uglifyjsplugintest to use fuse-box-test-runner still needs converting from createEnv
